+++ api/src/main/java/com/bookwhale/article/controller/ArticleController.java
@@ -1,17 +1,17 @@
 package com.bookwhale.article.controller;
 
-import com.bookwhale.article.dto.ArticlesResponse;
-import com.bookwhale.common.dto.Pagination;
-import com.bookwhale.article.dto.BookResponse;
-import com.bookwhale.common.dto.ConditionListResponse;
-import com.bookwhale.article.dto.NaverBookRequest;
 import com.bookwhale.article.dto.ArticleRequest;
 import com.bookwhale.article.dto.ArticleResponse;
 import com.bookwhale.article.dto.ArticleStatusUpdateRequest;
 import com.bookwhale.article.dto.ArticleUpdateRequest;
 import com.bookwhale.article.dto.ArticlesRequest;
-import com.bookwhale.article.service.NaverBookAPIService;
+import com.bookwhale.article.dto.ArticlesResponse;
+import com.bookwhale.article.dto.BookResponse;
+import com.bookwhale.article.dto.NaverBookRequest;
 import com.bookwhale.article.service.ArticleService;
+import com.bookwhale.article.service.NaverBookAPIService;
+import com.bookwhale.common.dto.ConditionListResponse;
+import com.bookwhale.common.dto.Pagination;
 import com.bookwhale.security.CurrentUser;
 import com.bookwhale.user.domain.User;
 import java.net.URI;
@@ -33,7 +33,7 @@ import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
 @RestController
-@RequestMapping("/api/article")
+@RequestMapping("/api")
 @RequiredArgsConstructor
 public class ArticleController {
 
@@ -41,14 +41,7 @@ public class ArticleController {
 
     private final NaverBookAPIService naverBookAPIService;
 
-    @GetMapping("/naverBookAPI")
-    public ResponseEntity<List<BookResponse>> findNaverBooks(
-        @Valid @ModelAttribute NaverBookRequest naverBookRequest) {
-        List<BookResponse> bookResponses = naverBookAPIService.getNaverBooks(naverBookRequest);
-        return ResponseEntity.ok(bookResponses);
-    }
-
-    @PostMapping
+    @PostMapping("/article")
     public ResponseEntity<Void> createArticle(@CurrentUser User user,
         @Valid @RequestPart("articleRequest") ArticleRequest articleRequest,
         @RequestPart(name = "images") List<MultipartFile> images) throws URISyntaxException {
@@ -56,19 +49,32 @@ public class ArticleController {
         return ResponseEntity.created(new URI("/api/article/" + articleId)).build();
     }
 
-    @GetMapping("/{articleId}")
+    @GetMapping("/article/naver-book")
+    public ResponseEntity<List<BookResponse>> findNaverBooks(
+        @Valid @ModelAttribute NaverBookRequest naverBookRequest) {
+        List<BookResponse> bookResponses = naverBookAPIService.getNaverBooks(naverBookRequest);
+        return ResponseEntity.ok(bookResponses);
+    }
+
+    @GetMapping("/articles/me")
+    public ResponseEntity<List<ArticlesResponse>> findMyArticles(@CurrentUser User user) {
+        List<ArticlesResponse> response = articleService.findMyArticles(user);
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping("/article/{articleId}")
     public ResponseEntity<ArticleResponse> findArticle(@CurrentUser User user,
         @PathVariable Long articleId) {
         return ResponseEntity.ok(articleService.findArticle(user, articleId));
     }
 
-    @GetMapping
+    @GetMapping("/articles")
     public ResponseEntity<List<ArticlesResponse>> findArticles(ArticlesRequest articlesRequest,
         @Valid Pagination pagination) {
         return ResponseEntity.ok(articleService.findArticles(articlesRequest, pagination));
     }
 
-    @PatchMapping("/{articleId}")
+    @PatchMapping("/article/{articleId}")
     public ResponseEntity<Void> updateArticle(@CurrentUser User user,
         @PathVariable Long articleId,
         @Valid @RequestPart("articleUpdateRequest") ArticleUpdateRequest request,
@@ -77,27 +83,26 @@ public class ArticleController {
         return ResponseEntity.ok().build();
     }
 
-    @PatchMapping("/articleStatus/{articleId}")
+    @PatchMapping("/article/{articleId}/status")
     public ResponseEntity<Void> updateArticleStatus(@CurrentUser User user, @PathVariable Long articleId,
         @Valid @RequestBody ArticleStatusUpdateRequest request) {
         articleService.updateArticleStatus(user, articleId, request);
         return ResponseEntity.ok().build();
     }
 
-    @DeleteMapping("/{articleId}")
+    @DeleteMapping("/article/{articleId}")
     public ResponseEntity<Void> deleteArticle(@CurrentUser User user, @PathVariable Long articleId) {
         articleService.deleteArticle(user, articleId);
         return ResponseEntity.ok().build();
     }
 
-    @GetMapping("/conditions/bookStatus")
+    @GetMapping("/article/conditions/bookStatus")
     public ResponseEntity<List<ConditionListResponse>> getAllBookStatus() {
         return ResponseEntity.ok(ConditionListResponse.listOfBookStatus());
     }
 
-    @GetMapping("/conditions/locations")
+    @GetMapping("/article/conditions/locations")
     public ResponseEntity<List<ConditionListResponse>> getAllSellingLocation() {
         return ResponseEntity.ok(ConditionListResponse.listOfSellingLocation());
     }
-
 }
