+++ api/src/test/java/com/bookwhale/article/docs/ArticleDocumentation.java
@@ -75,6 +75,33 @@ public class ArticleDocumentation {
             ));
     }
 
+    public static RestDocumentationResultHandler findMyArticles() {
+        FieldDescriptor[] response = new FieldDescriptor[]{
+            fieldWithPath("articleId").type(JsonFieldType.NUMBER).description("게시글 ID"),
+            fieldWithPath("articleTitle").type(JsonFieldType.STRING).description("게시글 제목"),
+            fieldWithPath("articlePrice").type(JsonFieldType.STRING).description("게시글 가격"),
+            fieldWithPath("sellingLocation").type(JsonFieldType.STRING).description(
+                "게시글에 등록한 판매지역"),
+            fieldWithPath("description").type(JsonFieldType.STRING).description("게시글에 작성한 설명"),
+            fieldWithPath("viewCount").type(JsonFieldType.NUMBER).description("게시글 조회수"),
+            fieldWithPath("favoriteCount").type(JsonFieldType.NUMBER).description("게시글 관심수"),
+            fieldWithPath("beforeTime").type(JsonFieldType.STRING).description("등록한 시간 - 현재 시간"),
+            fieldWithPath("articleImage").type(JsonFieldType.STRING).description("판매자가 올린 이미지"),
+            fieldWithPath("articleStatus").type(JsonFieldType.STRING).description(
+                "게시글 상태 [SALE, RESERVED, SOLD_OUT]")
+        };
+
+        return document("user/findMyArticles",
+            preprocessRequest(prettyPrint()),
+            preprocessResponse(prettyPrint()),
+            requestHeaders(
+                headerWithName(HttpHeaders.AUTHORIZATION).description("접속 인증 정보가 담긴 JWT")
+            ),
+            responseFields(fieldWithPath("[]").description("An arrays of articlesResponse"))
+                .andWithPrefix("[].", response)
+        );
+    }
+
     public static RestDocumentationResultHandler findArticle() {
         return document("article/findArticle",
             preprocessResponse(prettyPrint()),
