+++ api/src/test/java/com/bookwhale/user/acceptance/UserAcceptanceTest.java
@@ -231,66 +231,4 @@ public class UserAcceptanceTest extends AcceptanceTest {
         assertThat(favoriteRespons.size()).isEqualTo(0);
         assertThat(articleResponse.getFavoriteCount()).isEqualTo(favoriteCount - 1L);
     }
-
-    @DisplayName("나의 게시글들을 조회한다.")
-    @Test
-    void findArticles_one() {
-        ArticleRequest articleRequest = ArticleRequest.builder()
-            .bookRequest(BookRequest.builder()
-                .bookSummary("책 설명")
-                .bookPubDate("2021-12-12")
-                .bookIsbn("12345678910")
-                .bookListPrice("10000")
-                .bookThumbnail("썸네일")
-                .bookTitle("토비의 스프링")
-                .bookPublisher("허브출판사")
-                .bookAuthor("이일민")
-                .build())
-            .title("토비의 스프링 팝니다~")
-            .description("책 설명")
-            .sellingLocation("DAEGU")
-            .bookStatus("BEST")
-            .price("5000")
-            .build();
-
-        String jwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(loginRequest);
-        ArticleAcceptanceStep.requestToCreateArticle(jwt, articleRequest);
-
-        ExtractableResponse<Response> response = UserAcceptanceStep.requestToFindMyArticles(jwt);
-        List<ArticlesResponse> articlesRespons = response.jsonPath().getList(".", ArticlesResponse.class);
-
-        AcceptanceStep.assertThatStatusIsOk(response);
-        UserAcceptanceStep.assertThatFindMyArticles(articlesRespons, articleRequest);
-    }
-
-    @DisplayName("나의 게시글들을 조회할 때 다른 유저의 게시글을 조회되지 않는다.")
-    @Test
-    void findArticles_empty() {
-        ArticleRequest articleRequest = ArticleRequest.builder()
-            .bookRequest(BookRequest.builder()
-                .bookSummary("책 설명")
-                .bookPubDate("2021-12-12")
-                .bookIsbn("12345678910")
-                .bookListPrice("10000")
-                .bookThumbnail("썸네일")
-                .bookTitle("토비의 스프링")
-                .bookPublisher("허브출판사")
-                .bookAuthor("이일민")
-                .build())
-            .title("토비의 스프링 팝니다~")
-            .description("책 설명")
-            .bookStatus("BEST")
-            .price("5000")
-            .build();
-
-        String jwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(loginRequest);
-        String anotherJwt = UserAcceptanceStep.requestToLoginAndGetAccessToken(anotherLoginRequest);
-        ArticleAcceptanceStep.requestToCreateArticle(anotherJwt, articleRequest);
-
-        ExtractableResponse<Response> response = UserAcceptanceStep.requestToFindMyArticles(jwt);
-        List<ArticlesResponse> articlesRespons = response.jsonPath().getList(".", ArticlesResponse.class);
-
-        AcceptanceStep.assertThatStatusIsOk(response);
-        assertThat(articlesRespons.size()).isEqualTo(0);
-    }
 }
